pyproject.toml
[tool.poetry]
name = "chat-protector"
version = "0.1.0"
description = ""
authors = ["GreenFor3st <greenforest249@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.6"
vt-py = "^0.17.5"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
